modified basePage
