+++ src/api/api_popularlist.js
@@ -1,5 +1,4 @@
-import { baseUrl } from "./base";
-import { axiosInstance } from "./api.js"
+import { baseUrl, axiosInstance } from "./base";
 
   export const getPopularList = async() => {
     try{
